+++ src/main/java/com/pima/creditoparati/dao/services/ICategoryUserService.java
@@ -8,11 +8,11 @@ public interface ICategoryUserService {
 
     public List<CategoryUser> findAll();
 
-    public CategoryUser findById(Long id);
+    public CategoryUser findById(Integer id);
 
     public CategoryUser save(CategoryUser categoryUser);
 
-    public void delete(Long id);
+    public void delete(Integer id);
 
     public List<CategoryUser> findAllActive();
 }
